requirements.txt
pandas==2.0.0
numpy==1.24.3
pyyaml==6.0
matplotlib==3.7.1
seaborn==0.12.2
sqlalchemy==2.0.15